pyproject.toml
[project]
name = "guizhong"
version = "2.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2026.2.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
